Change in plugin filenames to pass tests.
