+++ as2_simulation_assets/as2_gazebo_assets/plugins/indi-control_ign/Common.cpp
