+++ as2_simulation_assets/as2_gazebo_assets/plugins/indi-control_ign/PIDController.hpp
@@ -59,8 +59,8 @@
  * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  ********************************************************************************/
 
-#ifndef AS2_SIMULATION_ASSETS__AS2_GAZEBO_ASSETS__PLUGINS__INDI_CONTROLLER__PIDCONTROLLER_HPP_
-#define AS2_SIMULATION_ASSETS__AS2_GAZEBO_ASSETS__PLUGINS__INDI_CONTROLLER__PIDCONTROLLER_HPP_
+#ifndef AS2_SIMULATION_ASSETS__AS2_GAZEBO_ASSETS__PLUGINS__INDI_CONTROL_IGN__PIDCONTROLLER_HPP_
+#define AS2_SIMULATION_ASSETS__AS2_GAZEBO_ASSETS__PLUGINS__INDI_CONTROL_IGN__PIDCONTROLLER_HPP_
 
 #include <math.h>
 #include <Eigen/Dense>
@@ -757,4 +757,4 @@ protected:
 
 }  // namespace pid_controller
 
-#endif  // AS2_SIMULATION_ASSETS__AS2_GAZEBO_ASSETS__PLUGINS__INDI_CONTROLLER__PIDCONTROLLER_HPP_
+#endif  // AS2_SIMULATION_ASSETS__AS2_GAZEBO_ASSETS__PLUGINS__INDI_CONTROL_IGN__PIDCONTROLLER_HPP_
